+++ clang/labs/lab1/alexseev_danila/deprecatedWarning.cpp
