+++ clang/labs/lab1/bodrov_daniil/ClassPrinter.cpp
