+++ clang/labs/lab1/durandin_vladimir/AlwaysInlineAttributePlugin.cpp
