+++ clang/labs/lab1/korablev_nikita/CMakeLists.txt
