+++ clang/labs/lab1/kostin_artem/depr_plugin.cpp
