+++ clang/labs/lab1/kulaev_zhenya/AttrubuteAlways.cpp
