+++ clang/labs/lab1/kulagin_aleksandr/alwaysInline.cpp
