+++ clang/labs/lab1/kulikov_artem/CMakeLists.txt
