+++ clang/labs/lab1/kulikov_artem/DeprecatedWarning.cpp
