+++ clang/labs/lab1/kuznetsov_artyom/CMakeLists.txt
