+++ clang/labs/lab1/petrov_maksim/CMakeLists.txt
