+++ clang/labs/lab1/petrov_maksim/WarningDeprecatedPlugin.cpp
