+++ clang/labs/lab1/polozov_vladislav/WarningForDeprecated.cpp
