+++ clang/labs/lab1/sadikov_damir/DeprecatedPlugin_SadikovD.cpp
