+++ clang/labs/lab1/sharapov_georgiy/CMakeLists.txt
