+++ clang/labs/lab1/sharapov_georgiy/DepWarningPlugin.cpp
