+++ clang/labs/lab1/soloninko_andrey/CMakeLists.txt
