+++ clang/labs/lab1/soloninko_andrey/classList.cpp
