+++ clang/labs/lab1/vasilev_ivan/CMakeLists.txt
