+++ clang/labs/lab1/vasilev_ivan/Find_Deprecation.cpp
