+++ clang/labs/lab1/volodin_evgeniy/PrintClassPlugin.cpp
